Restructure to workspace
