+++ .github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
           toolchain: stable
           components: clippy
       - name: Run Clippy
-        run: cargo clippy
+        run: cargo clippy -- -D warnings
       - name: Run Tests
         run: cargo test
