+++ .gitignore
@@ -1,5 +1,7 @@
 # Ignore build artifacts
 /target
+.dfx
+*.wasm
 
 # macOS metadata
 .DS_Store
