+++ Cargo.toml
@@ -1,16 +1,16 @@
 [package]
-name = "blockxpand-icp-aggregator"
+name = "blockxpand_icp"
 version = "0.1.0"
 edition = "2021"
+
 [dependencies]
-async-trait = "0.1"
-candid = "0.9"
-ic-agent = "0.26"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-ic-cdk = "0.10"
-ic-cdk-macros = "0.8"
-futures = "0.3"
+aggregator = { path = "src/aggregator" }
+bx_core = { path = "src/bx_core" }
 
 [dev-dependencies]
+candid = "0.9"
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
+
+[workspace]
+members = ["src/bx_core", "src/aggregator"]
+
