+++ src/aggregator/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "aggregator"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+async-trait = "0.1"
+candid = "0.9"
+ic-agent = "0.26"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+ic-cdk = "0.10"
+ic-cdk-macros = "0.8"
+futures = "0.3"
+bx_core = { path = "../bx_core" }
+
+[dev-dependencies]
+tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
