+++ src/aggregator/src/ledger_fetcher.rs
@@ -1,6 +1,8 @@
 use candid::Principal;
-pub async fn fetch(_principal: Principal) -> Vec<super::Holding> {
-    vec![super::Holding {
+use bx_core::Holding;
+
+pub async fn fetch(_principal: Principal) -> Vec<Holding> {
+    vec![Holding {
         source: "ledger".to_string(),
         token: "ICP".to_string(),
         amount: "213.45".to_string(),
