+++ src/bx_core/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "bx_core"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+candid = "0.9"
