+++ src/lib.rs
@@ -1,8 +1,2 @@
-pub mod aggregator;
-#[derive(serde::Serialize)]
-pub struct Holding {
-    pub source: String,
-    pub token: String,
-    pub amount: String,
-    pub status: String,
-}
+pub use aggregator::*;
+pub use bx_core::*;
